refactor: consolidate duplicated geometry primitives into core.geometry (#2349)
Move point_to_segment_distance, segment_to_segment_distance,
segments_intersect, and segment_clearance into a single canonical module
(core/geometry.py). Replace 15+ duplicate implementations across router,
validate, mcp, cli, and optim modules with imports from the canonical
source. Also consolidate extract_board_outline into placement/geometry.py
and optim/board_outline.py to eliminate duplicate sexp-parsing logic.
