+++ src/kicad_tools/validate/rules/clearance.py
@@ -10,6 +10,11 @@ import math
 from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
+from kicad_tools.core.geometry import (
+    point_to_segment_distance as _point_to_segment_distance,
+    segment_to_segment_distance as _segment_to_segment_distance,
+)
+
 from ..violations import DRCResults, DRCViolation
 from .base import DRC_TOLERANCE, DRCRule
 
@@ -151,43 +156,9 @@ def _transform_pad_dimensions(pad: Pad, footprint: Footprint) -> tuple[float, fl
     return new_width, new_height
 
 
-def _point_to_segment_distance(
-    px: float, py: float, x1: float, y1: float, x2: float, y2: float
-) -> float:
-    """Calculate the distance from a point to a line segment."""
-    # Vector from p1 to p2
-    dx = x2 - x1
-    dy = y2 - y1
 
-    # Length squared of segment
-    len_sq = dx * dx + dy * dy
-
-    if len_sq == 0:
-        # Segment is a point
-        return math.sqrt((px - x1) ** 2 + (py - y1) ** 2)
-
-    # Parameter t for the closest point on the line
-    t = max(0, min(1, ((px - x1) * dx + (py - y1) * dy) / len_sq))
-
-    # Closest point on segment
-    closest_x = x1 + t * dx
-    closest_y = y1 + t * dy
-
-    # Distance from point to closest point
-    return math.sqrt((px - closest_x) ** 2 + (py - closest_y) ** 2)
-
-
-def _segment_to_segment_distance(
-    x1: float, y1: float, x2: float, y2: float, x3: float, y3: float, x4: float, y4: float
-) -> float:
-    """Calculate minimum distance between two line segments."""
-    # Check all four endpoint-to-segment distances
-    d1 = _point_to_segment_distance(x1, y1, x3, y3, x4, y4)
-    d2 = _point_to_segment_distance(x2, y2, x3, y3, x4, y4)
-    d3 = _point_to_segment_distance(x3, y3, x1, y1, x2, y2)
-    d4 = _point_to_segment_distance(x4, y4, x1, y1, x2, y2)
-
-    return min(d1, d2, d3, d4)
+# _point_to_segment_distance and _segment_to_segment_distance are imported
+# from kicad_tools.core.geometry (consolidated in #2349).
 
 
 def _calculate_clearance(elem1: CopperElement, elem2: CopperElement) -> tuple[float, float, float]:
